fix: changed default fetch-limit in getInvoices to 1000
The reason for this change is that the api has changed to no longer accept a limit of -1.

At the same time, the function now sets the limit to 1000 when invalid limits like -1 are provided. Also when the limit is set to 0, then the function simply returns an empty array.